+++ Source/ilyvion.LoadingProgress/Constants.cs
@@ -1,0 +1,8 @@
+namespace ilyvion.LoadingProgress;
+
+public static class Constants
+{
+    public const float HorizontalMargin = 21f;
+    public const float VerticalWidgetMargin = 4f;
+    public const float ProgressBarHeight = 20f;
+}