+++ Source/ilyvion.LoadingProgress/LoadingProgressWindow.cs
@@ -1,14 +1,11 @@
 using System.Diagnostics;
 using UnityEngine;
+using static ilyvion.LoadingProgress.Constants;
 
 namespace ilyvion.LoadingProgress;
 
 public partial class LoadingProgressWindow
 {
-    private const float HorizontalMargin = 21f;
-    private const float VerticalWidgetMargin = 4f;
-    private const float ProgressBarHeight = 20f;
-
     private static readonly Vector2 BaseWindowSize = new(776f, 110f);
     internal static Vector2 WindowSize
     {
@@ -89,7 +86,7 @@ public partial class LoadingProgressWindow
         progressRect.height = ProgressBarHeight;
         if (StageProgress is (float currentValue, float maxValue))
         {
-            DrawHorizontalProgressBar(
+            Widgets_Progressbar.DrawHorizontalProgressBar(
                 progressRect,
                 (int)CurrentStage,
                 (int)LoadingStage.Finished,
@@ -98,7 +95,7 @@ public partial class LoadingProgressWindow
         }
         else
         {
-            DrawHorizontalProgressBar(
+            Widgets_Progressbar.DrawHorizontalProgressBar(
                 progressRect,
                 (int)CurrentStage,
                 (int)LoadingStage.Finished);
@@ -117,7 +114,7 @@ public partial class LoadingProgressWindow
             if (_lastLoadingTime.HasValue)
             {
                 float totalSeconds = (float)_lastLoadingTime.Value.TotalSeconds;
-                DrawHorizontalProgressBar(
+                Widgets_Progressbar.DrawHorizontalProgressBar(
                     loadingTimeRect,
                     Math.Clamp((float)elapsed.TotalSeconds, 0f, totalSeconds),
                     totalSeconds,
@@ -160,63 +157,6 @@ public partial class LoadingProgressWindow
         Text.Anchor = TextAnchor.UpperLeft;
     }
 
-    protected static void DrawHorizontalProgressBar(
-        Rect progressRect,
-        float currentValue,
-        float maxValue,
-        float? smallCurrentValue = null,
-        float? smallMaxValue = null,
-        Color? customBarColor = null,
-        Color? customSmallBarColor = null)
-    {
-        // draw a box for the bar
-        GUI.color = Color.gray;
-        Widgets.DrawBox(progressRect.ContractedBy(1f));
-        GUI.color = Color.white;
-
-        // get the bar rect
-        var barRect = progressRect.ContractedBy(2f);
-        var unit = barRect.width / maxValue;
-        barRect.width = currentValue * unit;
-
-        if (smallCurrentValue.HasValue && smallMaxValue.HasValue)
-        {
-            smallCurrentValue = smallCurrentValue.Value % (2 * smallMaxValue.Value);
-
-            // draw the small bar
-            var smallBarRect = progressRect.ContractedBy(2f);
-            smallBarRect.yMin += barRect.height / 2;
-            var smallUnit = smallBarRect.width / smallMaxValue.Value;
-            smallBarRect.width = smallCurrentValue.Value * smallUnit;
-
-            float clampedSmallCurrentValue = Mathf.Clamp(smallCurrentValue.Value, 0f, smallMaxValue.Value);
-            if (smallCurrentValue.Value > smallMaxValue.Value)
-            {
-                // Once we're past max, start drawing the bar with a gap on the left side equal to the overflow.
-                smallBarRect.width = smallMaxValue.Value * smallUnit;
-                smallBarRect.xMin += (smallCurrentValue.Value - smallMaxValue.Value) * smallUnit;
-            }
-
-            if (currentValue < maxValue)
-            {
-                // draw the big/main bar with "internal" progress
-                barRect.width += clampedSmallCurrentValue / smallMaxValue.Value * unit;
-            }
-            Widgets.DrawBoxSolid(barRect, customBarColor ?? BarColor);
-
-            // draw the small bar
-            Widgets.DrawBoxSolid(smallBarRect, customSmallBarColor ?? SmallBarColor);
-        }
-        else
-        {
-            // draw the big/main bar
-            Widgets.DrawBoxSolid(barRect, customBarColor ?? BarColor);
-        }
-    }
-
-    private static readonly Color BarColor = new(0.2f, 0.8f, 0.85f);
-    private static readonly Color SmallBarColor = Color.white.ToTransparent(0.75f);
-
-    private static readonly Color TimeBarColor = BarColor.Darken(0.2f);
+    private static readonly Color TimeBarColor = Widgets_Progressbar.BarColor.Darken(0.2f);
     private static readonly Color TimerSmallBarColor = Color.white.Darken(0.2f).ToTransparent(0.75f);
 }