+++ Source/ilyvion.LoadingProgress/ModContentPack_ReloadContentInt_Patch.cs
@@ -1,4 +1,9 @@
+using System.Diagnostics;
 using System.Reflection;
+
+using ilyvion.LoadingProgress.FasterGameLoading;
+
+using UnityEngine;
 
 namespace ilyvion.LoadingProgress;
 
@@ -6,6 +11,13 @@ namespace ilyvion.LoadingProgress;
 public static class ModContentPack_ReloadContentInt_Patch
 {
     internal static ModContentPack? CurrentModContentPack;
+
+    [HarmonyPrefix]
+    [HarmonyPriority(Priority.VeryHigh)]
+    public static void ProgressPrefix(ModContentPack __instance)
+    {
+        FasterGameLoadingProgressWindow.LoadingMod = __instance;
+    }
 
     [HarmonyPriority(Priority.VeryLow)]
     public static bool Prefix(ModContentPack __instance)
@@ -21,26 +33,5 @@ public static class ModContentPack_ReloadContentInt_Patch
             LoadingProgressMod.Debug("Letting original ReloadContentInt for " + __instance.Name + " run because we're not the one controlling it");
         }
         return shouldRunOriginal;
-    }
-}
-
-[HarmonyPatch]
-internal static class FasterGameLoading_DelayedActions_LateUpdate_Patches
-{
-    internal static bool _pauseFasterGameLoading_DelayedActions_LateUpdate;
-
-    private static bool Prepare()
-    {
-        return ModsConfig.ActiveModsInLoadOrder.Any(mod => mod.PackageId.Equals("taranchuk.fastergameloading", StringComparison.CurrentCultureIgnoreCase));
-    }
-
-    private static MethodInfo TargetMethod()
-    {
-        return AccessTools.Method("FasterGameLoading.DelayedActions:LateUpdate");
-    }
-
-    private static bool Prefix()
-    {
-        return !_pauseFasterGameLoading_DelayedActions_LateUpdate;
     }
 }